Test JWT-to-User parsing and correct type signatures
Also pulled out the JWKSSources to their own module for readability.
